move mode detect method to enum
